refactoring gui files to have one file to run everytyhing
