+++ src/main/java/gui/Label.java
@@ -1,0 +1,2 @@
+package gui;public class Label {
+}
