+++ src/main/java/gui/ScrabbleApp.java
@@ -1,0 +1,2 @@
+package gui;public class ScrabbleApp {
+}
